:bad: Change concept, convert spaceX to anime
Commented spaceX code in main.js, this has to be removed.

in render.js converted render function to render the anime list in a dirty way. Currently it's doing an if loop to see if an english title exists or it should return an english-japanese title. Has to be split.